ensure consistent thickness and tracer budgets on diagnostic grids, part 2
Refactor tracer module column physics subroutine to perform a single action
in each invocation. Actions are one of
  col_act_apply_KPP_NLT
  col_act_pre_tridiag_solve_sources
  col_act_apply_boundary_fluxes
  col_act_tridiag_solve
  col_act_post_tridiag_solve_sources
This enables each action to be performed immediately after that action is
performed for t and s in the diabatic driver.

Generalize tendency diagnostics for following processes to passive tracers:
  frazil formation
  KPP non-local transport
  boundary forcing
  diabatic mixing

If KPP non-local transport tendencies are requested, the term is applied
independent of applyNonLocalTrans. If applyNonLocalTrans==.false., the
application is undone after diagnostics are posted.

Change intent of h argument of applyTracerBoundaryFluxesInOut to intent(in),
and add optional h_new intent(out) argument for updated thicknesses.
The only application needing the updated thickness is offline mode.

use case consistently for tracer_Reg in MOM.F90