Add Leith+E
This commit adds the 2D Leith+E closure, which uses a modified
2D Leith biharmonic viscosity paired with a harmonic backscatter.
('Modified' here is not used in the same sense as 'modified
2D Leith'; it just means that the biharmonic coefficient is
modified to account for enstrophy backscatter.) Variables
are often named 'leithy' to refer to Leith+E.

The parameterization is controlled by three main entries in
user_nl_mom:
1. USE_LEITHY = True
2. LEITH_CK = 1.0
3. LEITH_BI_CONST = 8.0

To use Leith+E you should have LAPLACIAN=True and BIHARMONIC=True.
(It doesn't hurt to be explicit and also set LEITH_AH=False, along
with any other viscous closures, but this is not required. If
USE_LEITHY=True it will not use any of the other schemes.)
LEITH_CK is the fraction of energy dissipated by the biharmonic
term that gets backscattered by the harmonic term (it's a target;
the backscatter rate is not exact.) Recommended values between 0
and 1.
LEITH_BI_CONST is Upsilon^6 where Upsilon is the ratio between
the dissipation scale for enstrophy and the grid scale. Values
should probably be greater than or equal to 1.

This commit also includes a merge with dev/ncar. Leith+E only
changes MOM_hor_visc.F90; all other changes are associated with
the merge.