Scale checksums in hor_bnd_diffusion
  Added missing scale arguments to the hchksum and global_mass_integral calls
for debugging in hor_bnd_diffusion, so that they now give messages to stdout
that do not change when tracers (including temperature and salinity) are
rescaled.  Also added a missing debuggingParam argument to the get_Param call
for HBD_DEBUG so that is will be logged in MOM_parameter_doc.debugging rather
that MOM_parameter_doc.all.  This commit partially addresses the scaling
problems that were noted in github.com/NCAR/MOM6/issues/275.  All solutions are
bitwise identical, but some debugging output can change to become more robust.