Add option to auto-compute an io layout for auto masking.
When Auto Masking is enabled, the LAYOUT is unknown a priori,
so it's not possible to predetermine a compatible IO_LAYOUT.
This commit introduces an option to set a target number of
IO PEs (TARGET_IO_PES), to be used as a reference for the
auto masking routine to come up with an IO_LAYOUT compatible
with LAYOUT on the fly. If original TARGET_IO_PES is not
achievable without having to sacrifice too much land masking,
IO PEs is decremented until a feasible combination of LAYOUT
and IO_LAYOUT is found.